refactor: move test contexts to tools
